Refactoring timeline events - split functionallity into several modules
